requirements.txt
requests==2.25.1
pytz==2021.1
dataclasses
discord==1.7.3
python-dotenv==0.18.0
git+https://github.com/Rapptz/discord.py
git+https://github.com/redParrot17/discord-ext-fancy-help
git+https://github.com/Rapptz/discord-ext-menus
